Draw paddles and rought playground
The paddles are controlled by the same keyboard: up, down, w, s keys are
used.

There are still graphical improvements to be done to make sure
paddles are not going through the "fence" when they get too fast.

Probably a terminal velocity is also needed - need to read on that
physics thing.

Everything is very crude but operational.